fix: tipo do atributo quantidade

+++ Codigo/Back/src/main/java/com/hortifruti/sl/hortifruti/dto/purchase/GroupedProductResponse.java
@@ -3,4 +3,4 @@ package com.hortifruti.sl.hortifruti.dto.purchase;
 import java.math.BigDecimal;
 
 public record GroupedProductResponse(
-    String code, String name, BigDecimal price, Integer quantity, BigDecimal totalValue) {}
+    String code, String name, BigDecimal price, BigDecimal quantity, BigDecimal totalValue) {}
